+++ lib/hotwire_native_version_gate/concern.rb
@@ -9,8 +9,12 @@ module HotwireNativeVersionGate
         VersionGate.native_feature(feature, ios: ios, android: android)
       end
 
-      def native_version_regex=(regex)
-        VersionGate.native_version_regex = regex
+      def native_version_regexes
+        VersionGate.native_version_regexes
+      end
+
+      def native_version_regexes=(regexes)
+        VersionGate.native_version_regexes = regexes
       end
     end